requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.9
pydantic==2.11.3
pydantic-settings==2.8.1
sqlalchemy==2.0.40
anthropic==0.51.0
openai>=1.0.0
chromadb==0.5.23
sentence-transformers==3.4.1
pypdf==5.4.0
python-docx==1.1.2
orjson==3.10.16
streamlit==1.44.0
pandas==2.2.3
